When using LuaJIT, prefer actual package names in BRs
compat-lua-lpeg and compat-lua-mpack are provided by lua5.1-lpeg and lua5.1-mpack respectively.

On Fedora they share source RPMs with the main lpeg and mpack binary RPMs (thus lua-lpeg and lua-mpack).

On EPEL8, though, lua-lpeg is part of RHEL so there is a separate lua5.1-lpeg, and the maintainer does not think having it provide compat-lua-lpeg is a good idea, and recommended neovim depend on lua5.1-lpeg directly - see https://src.fedoraproject.org/rpms/lua5.1-lpeg/pull-request/1

neovim's new EPEL8 branch already uses these BRs, and it'll be nice to keep the delta as small as possible.

https://src.fedoraproject.org/rpms/neovim/blob/epel8/f/neovim.spec